Refactored BeeBaseDB and HymBaseDB classes, added test coverage
