+++ testdata/fasta/hymbase-format-Sihb-prot.fa
@@ -1,0 +1,38 @@
+>SINV12975-PA gnl|Sinv_1.0|SINV12975-PA
+PELQVNGNGVLSFARAMQRFFNIAFPLDDPVIAPLYTHVDTRASGTIYWGETDSPEVLSR
+AGGLIRSTFKNANTFVPTHVFLATWLDVGYYNERNDKNINNVHRVVFQVNTYQVAISSNG
+THSFVELLYPENGIQWIQGESHPNGLPDAKAQAGFMSEGKMYTLRGSGTDQIQNVDKWSN
+VNRPGQWLFQVGPIDEGENINTPDNVDDVILTMLIELSNITLLDTMTMHQAASCRSGATT
+CHSKATCIDYEIGFCCSCKQGYFGNGKSCQPNDVPLRVIGRISGTLNGIEFPSRDLQCYV
+QTKDGRTYTALSRVPEEIGASFQLLNNLGGVIGWLFAKPVSDTENGYELTGKKLMLYVFT
+YRNLRNRLCLLNIVSCCRRKAGLRLRRYXYVRVSRIGGLFNHTAELNFPASGDKVTIHSN
+YLGLDVFGQLKMEANITGTLPHLEQDTRVEYGDYDELYTRAQPGLIRSQSERTCKLRSGG
+NERELNFKLDQMINYQECPHATFDPEDDTTRLKYFRGVTAYESTEGIIRFAVNTKTTPLE
+EEDPCIQGRQTCGAHSSCVVDGDSFRCVCNAGYQYLYEEDGSAICVDVNECTAGNHMCSP
+DAQCINQEGSHMCQCRSGFTGDGRVCERLAACEDTRCGNYEQCAMIGGAPTCICMPGFEE
+TDQGCYPSQHGKIXFHISNKXINRVFLLIT
+>SINV22505-PA gnl|Sinv_1.0|SINV22505-PA
+VPFKIVSLGSYTATEALFSVFVAVLKDFYWKTFELVGYAFLDILDNREIVHKEFVPPGQT
+VNQTFYREVVKRLRKKMALPVTFFILPAQKPFKRASLCTLYNIQKSVIDKRKGVPAEAFQ
+NCYEDWKQRLRHCVTAQGHYFKGNNFEV
+>SINV18683-PA gnl|Sinv_1.0|SINV18683-PA
+MCNVNNRTTSNMKMRVLVGVFLACGVVAGLPVQDEGLPNNILTPLDYALFENKEAYVYDD
+NMNLVKLMWDDAEEVGDQLNDSEIPSRVFFYLYKKNDPNPKELYVDNEDVLKNSNFDPAK
+PTRFVTHGWINSRNSEACTLVRDGYLQHDDYNVIVVDWSSISMRPYIWASNHVVPIARFV
+ATMINFLVKHGMNPSQTILVGHSLGAHVVGIAARNANSDIGYVVGLDPALPNFHLAGPGS
+RISSGDAKYVEIIHTNGGLLGFLVPIGDVDFYPNGGRKQLGCIVDAGGACSHARSYRFFA
+ESINSKVGFHGKSCSSYARFKLGLCKNGHTSIMGGHKPLLSARGNYYLMTNPSAPYASG
+>SINV20618-PA gnl|Sinv_1.0|SINV20618-PA
+LLIPGQHIRPPVTEETAIALLERLYGMRATSVRELNAYDDRNYHVVCGGTTTNPHAINLE
+NTSYVLKVINSLESRKTDVIEAQTELMIFLNQQNICCSLPVKNVSGAYFSLETLKSENTM
+EKYAVRLLVYRPGELLSRVKITDELLCKLGSFTARIDEILMGFNHPAYDNHKSLWMLTSV
+PQVQQFTYAIKSSFDKQLVHDVIVNFQKKVLPVTDQLEQTVIHGDLNEHNIIVTPDGKDI
+AAVIDFGDSHRTCRIFELAIVLCYMILQTADVGMGRYVVEGYEKIKKLTEMEKKILKISV
+CARICQSLVMGAYSHLHDPENQYLLSTQKTGWLLLKRLWPMSQEAVMQAWGL
+>SINV17869-PA gnl|Sinv_1.0|SINV17869-PA
+EARFREISTTVVEPTFAFFRPRRMDNFIKIEKIGEGTYGVVYKGKHKKTGEIVAMKKIRL
+ESDDEGIPSTAIREISLLKELTHPNIVSLIDVLMEESKLYLIFEYLTMDLKKYMDTLGNR
+MMEPAVVKSYLYQITRAILFCHKRRILHRDLKPQNLLIDKTGIIKVADFGLGRAFGIPVR
+IYTHEVVTLWYRAPEILLGATRYSCAIDMWSIGCIFSEMVTKKPLFQGDSEIDQLFRIFR
+ILRTPTEDIWPGVTQLSDYKATFPNWMTNNLESQVKTLDADGLDLLQAMLTYDPVYRISA
+RAALQHPYFSDLDTCNIPAA
